:hammer: Fix: table 필드 수정으로 인한 엔티티 수정

See merge request s08-webmobile1-sub2/S08P12A702!102

+++ back/Dream-Hi/src/main/java/com/elephant/dreamhi/model/dto/ActorListResponseDto.java
@@ -21,7 +21,7 @@ public class ActorListResponseDto {
     private String title;
     private Gender gender;
     private Integer age;
-    private Double height;
+    private Integer height;
     private List<StyleDto> styles;
     @NotNull
     private Boolean isFollow;
